Use proper feature test for program_invocation_short_name
Previously we were using __linux__ to test for the extension which
enabled program_invocation_short_name, but it is actually a GNU
extension. It would not be present if we were to link ulibc instead of
glibc.

This patch makes it so we instead use __GNUC__ as an indicator. This is
more correct.