Fix relative includes with -I
Since we're passing the source file to CC over stdin, #include paths are
resolved relative to CWD rather the location of the source file.

This is pretty unintuitive, so to mitigate it, this patch adds the
directory of the source file to the include path with -I.

I also use i++ pattern to avoid fucking up.