Run different-language test as part of """testsuite"""
This patch runs the CPP example as part of the `make test` target. With
this, all of the examples should now be run when running `make test` or
`make installtest`.

Since this example is only intended to demonstrate compilation to other
targets, I also removed the use of `/usr/bin/env`.